+++ CHANGELOG.md
@@ -2,10 +2,18 @@
 
 All notable changes to the Commands Explorer extension will be documented in this file.
 
+## 1.1.0
+
+- Introduced a new feature to mark commands as favorites
+- Added a new setting to display favorite commands in both their original groups and the favorites group (default: true)
+- Added a new setting to sync the favorites between devices (default: true)
+- Included a new command to directly open the Commands Explorer settings
+- Enhanced the organization of commands and command groups with alphabetical sorting
+
 ## 1.0.3 to 1.0.4
 
 - Switched to round logo
-- UPdated round logo
+- Updated round logo
 
 ## 1.0.0 to 1.0.2